requirements.txt
asgiref==3.8.1
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
Django==4.2.20
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
fsspec==2025.7.0
gunicorn==23.0.0
huggingface-hub==0.34.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
pillow==11.3.0
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.54.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
